refactor(cli): remove 'run' command and fold behavior into 'boot'\n\n- Delete RunCommand (not a simctl verb)\n- Allow 'xsim boot' with no args to boot last used when none running\n- Remove registration and update README
