+++ Sources/XSimCLI/Commands/StartCommand.swift
@@ -8,14 +8,16 @@ class StartCommand: BaseSimCommand, Command {
     let longDescription = """
     Boots the specified simulator device.
     You can specify the device by name or UUID.
+    If no device is specified and no simulators are running, boots the most recently used simulator.
 
     Examples:
+      xsim boot                                # boot last used (if none running)
       xsim boot "iPhone 16"                    # by name
       xsim boot 12345678-1234-1234-1234-123456789012  # by UUID
       xsim boot "iPhone 16" --runtime "iOS 26"   # disambiguate by runtime
     """
 
-    @Param var deviceIdentifier: String
+    @Param var deviceIdentifier: String?
 
     @Key("--runtime", description: "Filter by runtime when selecting by name (e.g. 'iOS 17', '17.0' or a runtime identifier)")
     var runtimeFilter: String?
@@ -24,18 +26,49 @@ class StartCommand: BaseSimCommand, Command {
 
     func execute() throws {
         do {
-            stdout <<< "Booting simulator...".dim
-
-            // Start the simulator
             let simulatorService = try getService()
-            try simulatorService.startSimulator(identifier: deviceIdentifier, runtimeFilter: runtimeFilter)
 
-            // Get device info for confirmation
-            let devices = try simulatorService.listDevices()
-            if let device = findDevice(in: devices, identifier: deviceIdentifier) {
-                displayStartSuccess(device: device)
+            if let identifier = deviceIdentifier {
+                stdout <<< "Booting simulator...".dim
+                // Boot the specified simulator
+                try simulatorService.startSimulator(identifier: identifier, runtimeFilter: runtimeFilter)
+
+                // Get device info for confirmation
+                let devices = try simulatorService.listDevices()
+                if let device = findDevice(in: devices, identifier: identifier) {
+                    displayStartSuccess(device: device)
+                } else {
+                    stdout <<< "✓ Booted simulator '\(identifier)'".green
+                }
             } else {
-                stdout <<< "✓ Booted simulator '\(deviceIdentifier)'".green
+                // No identifier specified: if nothing is running, boot the most recently used simulator
+                let devices = try simulatorService.listDevices()
+                let running = devices.filter(\.state.isRunning)
+                if !running.isEmpty {
+                    stdout <<< "Simulators already running (\(running.count)). Use 'xsim list --running' to view.".yellow
+                    return
+                }
+
+                if let lastUDID = RecentDeviceStore.lastBootedUDID() {
+                    if let dev = devices.first(where: { $0.udid == lastUDID }) {
+                        let typeName = DisplayFormat.deviceTypeName(from: dev.deviceTypeIdentifier)
+                        let runtimeName = DisplayFormat.runtimeName(from: dev.runtimeIdentifier)
+                        stdout <<< "Booting last used simulator: \(dev.name) (\(typeName), \(runtimeName))".dim
+                    } else {
+                        stdout <<< "Booting last used simulator...".dim
+                    }
+                    try simulatorService.startSimulator(identifier: lastUDID)
+                    // Show info after boot
+                    let updated = try simulatorService.listDevices()
+                    if let device = updated.first(where: { $0.udid == lastUDID }) {
+                        displayStartSuccess(device: device)
+                    } else {
+                        stdout <<< "✓ Simulator booted".green
+                    }
+                } else {
+                    stdout <<< "No recent simulator recorded.".yellow
+                    stdout <<< "Specify a device: xsim boot <name|UUID> or list: xsim list".dim
+                }
             }
 
         } catch let error as SimulatorError {
