+++ Sources/XSimCLI/XSimCLI.swift
@@ -27,7 +27,6 @@ public class XSimCLI {
     private func registerCommands() {
         // Register all available commands without heavy initialization
         cli.commands = [
-            RunCommand(),
             ListCommand(),
             StartCommand(), // boot
             StopCommand(), // shutdown
